+++ app/src/main/java/io/github/plenglin/vibratingclock/VibrationInterval.java
@@ -5,19 +5,19 @@ public class VibrationInterval {
     /**
      * The vibration pattern to play
      */
-    private int[] pattern;
+    private long[] pattern;
 
     /**
      * The interval between happenings
      */
     private long interval;
 
-    public VibrationInterval(int[] pattern, long interval) {
+    public VibrationInterval(long[] pattern, long interval) {
         this.pattern = pattern;
         this.interval = interval;
     }
 
-    public int[] getPattern() {
+    public long[] getPattern() {
         return pattern;
     }
